More bug fixes for 1.0.1

+++ tests/test_data.py
@@ -175,7 +175,7 @@ def test_spect_data_write_hyp(temp_dir, device, sos, eos):
     os.makedirs(feat_dir)
     torch.save(torch.rand(3, 3), os.path.join(feat_dir, 'a.pt'))
     data_set = data.SpectDataSet(temp_dir, sos=sos, eos=eos)
-    z = torch.randint(10, (4, 3), dtype=float)
+    z = torch.randint(10, (4, 3), dtype=torch.float)
     zz = z
     if sos:
         zz = torch.cat([torch.full_like(zz, sos), zz])
